nvidia-jetson-nano/samples/iot-edge/NvidiaJetsonEdgeSolution/modules/RequestProcessorModule/requirements.txt
azure-iot-device~=2.0.0
azure-storage-blob~=12.3.2